+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.97 (Mar 11, 2023)
+* Changed `nullstone status` display format for times to emit local times using "Mon Jan _2 15:04:05 MST". 
+* Changed `nullstone logs` display format for times to emit local times using "Mon Jan _2 15:04:05 MST". 
+
 # 0.0.96 (Feb 24, 2023)
 * Added `Variable.HasValue` helper function when running capability generation (`nullstone workspaces select`).